requirements.txt
certifi==2025.6.15
clickhouse-connect==0.8.18
kafka-python==2.2.14
lz4==4.4.4
pytz==2025.2
urllib3==2.5.0
zstandard==0.23.0
